+++ configuration.properties
@@ -42,3 +42,5 @@ merchantuser_team3=Team3Merchant
 merchantPassword=Flavor.2106
 
 
+
+
